Almost made map saving and loading - 1 minor fix to go
Need to fix default qtablewidgetitem saving issue, because it shows as "nullptr" if you take it with "item(i, j)" function. So i need to create all of them manually or make something else to solve it. Then i need to check program for issues with data overriding in the memory of the running program when i save/write not only once in a session